added recipe category choosing page under /recipes and all recipes under dynamic route which catches matching routes
