+++ src/components/recipes/category-card.tsx
@@ -1,0 +1,5 @@
+export default function CategoryCard() {
+  return (
+    <div className="rounded-xl border-2 border-black  shadow-button duration-300 hover:translate-x-[2px] hover:translate-y-[-2px]"></div>
+  );
+}
