+++ src/types.ts
@@ -1,0 +1,19 @@
+export type RecipeCategory =
+  | "all"
+  | "breakfast"
+  | "lunch"
+  | "dinner"
+  | "appetizers"
+  | "drinks"
+  | "desserts";
+
+export type RecipeCategoryData = Record<
+  RecipeCategory,
+  {
+    name: string;
+    titleText: string;
+    subtitleText: string;
+    href: string;
+    imageSize?: number;
+  }
+>;
